requirements.txt
Flask
opencv-python
pandas
ultralytics
torch
torchvision
Pillow
numpy
dataloader
inference